+++ ember-cli-build.js
@@ -6,7 +6,10 @@ const EmberAddon = require('ember-cli/lib/broccoli/ember-addon');
 
 module.exports = function (defaults) {
   let app = new EmberAddon(defaults, {
-    // Add options here
+    // config for ember-cli-babel
+    'ember-cli-babel': {
+      includePolyfill: true
+    },
   });
 
   /*